rename normalize_cuda -> normalize_8
The _cuda suffix is not included in the kernels.hpp api as for now,
although it may be nice. Anyways, then name normalize_cuda is used
also for template in kernels.cu so rename it. The suffix _8 semms to be
appropriate, because it indicates that it works in 8 bit (in).