kernels rgb->yuv (for WebP): fix s/width/height/
the MIN there is to check overflow when odd number of lines
is present but in that case, the value of pixel on row (width-1) was
taken. in case of tile in T33UUS where width < height (11720x11298),
the last few lines were just duplicied but if with were bigger
than height, it could have result in out-of-memory access for the
last odd line.