norotate: respect
if delegates were run (should be always now), the norotate option was ignored

Turrent implementation skips rotation but keeps setting bounds (needed for
tiles) - this is perhaps unneeded, tilling and norotate may be exclusive
and dhr_visualizer should not use norotate. so `return take_ownership` should
be ok as well.